Finished front-end of login
